# src/router/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telicent ACCESS",
    "version": "0.11.2",
    "description": "Telicent ACCESS API",
    "contact": {
      "name": "Telicent",
      "url": "https://telicent.io",
      "email": "info@telicent.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8091"
    }
  ],
  "paths": {
...
    "/scim/v2/Users": {
      "get": {
        "summary": "Get all SCIM users",
        "tags": [
          "SCIM"
        ],
        "description": "Returns all SCIM users",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved array of SCIM users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMUsersListResponse"
                }
              }
            }
          },
          "405": {
            "description": "SCIM is not enabled, method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMNotEnabled"
                }
              }
            }
          },
          "422": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMInvalid"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new SCIM user",
        "tags": [
          "SCIM"
        ],
        "description": "Create a new SCIM user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSCIMUser"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Successfully created SCIM user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredSCIMUser"
                }
              }
            }
          },
          "405": {
            "description": "SCIM is not enabled, method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMNotEnabled"
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMUserExists"
                }
              }
            }
          },
          "422": {
            "description": "User object invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMInvalid"
                },
                "example": {
                  "code": 422,
                  "message": "User object invalid"
                }
              }
            }
          }
        }
      }
    },
    "/scim/v2/Users/{userId}": {
      "get": {
        "summary": "Get specific SCIM user by UUID",
        "tags": [
          "SCIM"
        ],
        "description": "Returns specific SCIM user based on specified UUID",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user id to retrieve attributes for"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved SCIM user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredSCIMUser"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMUserNotFound"
                }
              }
            }
          },
          "405": {
            "description": "SCIM is not enabled, method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMNotEnabled"
                }
              }
            }
          },
          "422": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMInvalid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an existing SCIM user",
        "tags": [
          "SCIM"
        ],
        "description": "Deletes an existing SCIM user based on specified user ID",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user identifier for user to be deleted"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted existing SCIM user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSCIMUserSuccess"
                }
              }
            }
          },
          "405": {
            "description": "SCIM is not enabled, method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMNotEnabled"
                }
              }
            }
          },
          "422": {
            "description": "User deactivation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMInvalid"
                },
                "example": {
                  "code": 422,
                  "message": "User deactivation failed"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch an existing SCIM user",
        "tags": [
          "SCIM"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user identifier for user to be patched"
          }
        ],
        "requestBody": {
          "description": "Changes to user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Operations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "op": {
                          "type": "string",
                          "example": "replace"
                        },
                        "value": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Patch an existing SCIM user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "501": {
            "description": "Operation in patch not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMUnableToPatch"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Patch an existing SCIM user via PUT",
        "tags": [
          "SCIM"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user identifier for user to be patched"
          }
        ],
        "requestBody": {
          "description": "Changes to user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Operations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "op": {
                          "type": "string",
                          "example": "replace"
                        },
                        "value": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Patch an existing SCIM user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "501": {
            "description": "Operation in patch not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SCIMUnableToPatch"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "description": "Returns all users",
        "produces": [
          {
            "application/json": null
          }
        ],
        "responses": {
          "200": {
            "description": "An array of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvalid"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "description": "Creates a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created user ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserSuccess"
                }
              }
            }
          },
          "405": {
            "description": "SCIM currently enabled, create not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnableToCreateUser"
                }
              }
            }
          },
          "409": {
            "description": "User already exists in IdP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserExists"
                }
              }
            }
          },
          "422": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvalid"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ],
        "description": "Returns user based on specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "422": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvalid"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update an existing user",
        "tags": [
          "Users"
        ],
        "description": "Updates an existing user",
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully updated an existing user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Failed to update user - invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvalidRequest"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "422": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvalid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete existing user",
        "tags": [
          "Users"
        ],
        "description": "Deletes an existing user by specified ID (if SCIM is enabled, delete is a soft delete)",
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted existing user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserSuccess"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "422": {
            "description": "User delete failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvalid"
                },
                "example": {
                  "code": 422,
                  "message": "User delete failed"
                }
              }
            }
          }
        }
      }
    },
    "/users/lookup/{email}": {
      "get": {
        "summary": "Get user attributes by email",
        "tags": [
          "Users"
        ],
        "description": "Returns user attributes based on specified email address",
        "parameters": [
          {
            "in": "path",
            "name": "email"
          }
        ],
        "producers": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "6564acddf66ddafd4a2fd645"
                    },
                    "attributes": {
                      "$ref": "#/components/schemas/AttributesToStrings"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "422": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvalid"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SCIMUser"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "description": "Identifier for the user in Telicent ACCESS",
                "type": "string",
                "example": "507f1f77bcf86cd799439011"
              },
              "externalId": {
                "description": "Auth identifier for the user in Telicent ACCESS",
                "type": "string",
                "example": "507f1f77bcf86cd799439011"
              },
              "name": {
                "type": "string",
                "example": "User"
              },
              "userName": {
                "type": "string",
                "example": "User"
              },
              "email": {
                "type": "string",
                "example": "newuser@telicent.io"
              },
              "labels": {
                "type": "array",
                "example": [
                  {
                    "name": "nationality",
                    "value": "GBR",
                    "toString": "nationality=\"GBR\"",
                    "toDataLabelString": "permittedNationalities='GBR'",
                    "_id": "6540c57326bc7a0bd6658494"
                  },
                  {
                    "name": "clearance",
                    "value": "S",
                    "toString": "clearance=\"S\"",
                    "toDataLabelString": "classification='S'",
                    "_id": "6540c57326bc7a0bd6658495"
                  },
                  {
                    "name": "personnelType",
                    "value": "GOV",
                    "toString": "personnelType=\"GOV\"",
                    "toDataLabelString": null,
                    "_id": "6540c57326bc7a0bd6658496"
                  },
                  {
                    "name": "deployedOrganisation",
                    "value": "Org1",
                    "toString": "deployedOrganisation=\"Org1\"",
                    "toDataLabelString": "permittedOrganisations='Org1'",
                    "_id": "6540c57326bc7a0bd6658497"
                  }
                ]
              },
              "active": {
                "type": "boolean",
                "example": false
              },
              "userGroups": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "urn:telicent:groups:developers",
                  "urn:telicent:groups:group2"
                ]
              },
              "schemas": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "urn:ietf:params:scim:schemas:core:2.0:User"
                ]
              }
            }
          }
        ]
      },
      "CreateUserInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "User"
          },
          "email": {
            "type": "string",
            "example": "newuser@telicent.io"
          },
          "clearance": {
            "type": "string",
            "enum": [
              "O",
              "OS",
              "S",
              "TS"
            ]
          },
          "deployedOrganisation": {
            "type": "string",
            "example": "Organisation name"
          },
          "nationality": {
            "type": "string",
            "example": "GBR"
          },
          "personnelType": {
            "type": "string",
            "enum": [
              "GOV",
              "NON-GOV"
            ]
          },
          "active": {
            "type": "boolean",
            "example": false
          },
          "userGroups": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "urn:telicent:groups:developers",
              "urn:telicent:groups:group2"
            ]
          }
        }
      },
      "CreateUserSuccess": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          }
        }
      },
...
      "DeleteUserSuccess": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "UserInvalid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 422
          },
          "message": {
            "type": "string",
            "example": "Operation failed / User object invalid / Operator not supported"
          }
        }
      },
...
      "ServerError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 500
          },
          "message": {
            "type": "string",
            "example": "Server error"
          }
        }
      },
...
      "StoredSCIMUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SCIMUser"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Access identifier for user",
                "example": "64dd02xxxxxxxxxxxcecbe13c"
              }
            }
          }
        ]
      },
      "SCIMUsersListResponse": {
        "type": "object",
        "properties": {
          "schemas": {
            "type": "array",
            "items": "string",
            "example": [
              "urn:ietf:params:scim:schemas:core:2.0:User"
            ]
          },
          "totalResults": {
            "type": "number"
          },
          "startIndex": {
            "type": "number"
          },
          "itemsPerPage": {
            "type": "number"
          },
          "Resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoredSCIMUser"
            }
          }
        }
      },
      "CreateSCIMUser": {
        "type": "object",
        "required": [
          "externalId"
        ],
        "properties": {
          "externalId": {
            "type": "string",
            "example": "31127653-a234-48bc-b940-34c866bfe837"
          },
          "userName": {
            "type": "string",
            "example": "User"
          },
          "emails": {
            "type": "array",
            "example": [
              {
                "value": "user@telicent.io"
              }
            ]
          },
          "schemas": {
            "type": "array",
            "example": [
              "urn:ietf:params:scim:schemas:core:2.0:User"
            ]
          },
          "groups": {
            "type": "array",
            "example": [
              "urn:telicent:groups:developers",
              "urn:telicent:groups:group2"
            ]
          }
        }
      },
      "DeleteSCIMUserSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "ok": {
                "type": "boolean",
                "example": true
              }
            }
          }
        }
      },
      "SCIMUserNotFound": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "SCIM user not found"
          }
        }
      },
      "SCIMNotEnabled": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 405
          },
          "message": {
            "type": "string",
            "example": "SCIM is not enabled on this server"
          }
        }
      },
      "SCIMUserExists": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "SCIM user already exists"
          }
        }
      },
      "SCIMInvalid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 422
          },
          "message": {
            "type": "string",
            "example": "Operation failed / SCIM user object invalid / Operator not supported"
          }
        }
      },
      "SCIMUnableToPatch": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 501
          },
          "message": {
            "type": "string",
            "example": "Operation in patch not supported, only user deactivation is supported from the SCIM Service provider"
          }
        }
      },
...
    }
  },
...
}